move validations to shared/utils
