+++ app/javascript/packs/shared/utils/validations.js
