+++ app/javascript/staff/components/organization/form.vue
@@ -66,7 +66,7 @@
 </template>
 
 <script>
-import { validationRules } from "../../utils/validations";
+import { validationRules } from "../../../packs/shared/utils/validations"
 import { dadataClient } from "../../utils/dadataClient";
 
 export default {
